sound_Card_init: Fix the logic of invalid shutdown time
When the shutdown time file is invalid we assume the
speakers are overheat so that the boot time
calibration is not triggered and the previous calibration
results will be used.

The shutdown time can be invalid if `/var/lib/cras/stop`
is not updated when cras terminates like sudden power off.
With high possibility that the previous calibration result
should be used because users rarely replace their speakers.

BUG=b:218458902
TEST=`/usr/bin/sound_card_init --id=${SOUND_CARD_ID}
     --conf=${CONFIG} --amp=${AMP}` under invalid /var/lib/cras/stop and
      verify the applied calibration values matches the datastore values.

Change-Id: Iaab08d99b6b11ed4a5e6145d4bba9cd79c8fca2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3483048
Tested-by: Judy Hsiao <judyhsiao@google.com>
Auto-Submit: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>