Floss: memset the bytes field instead of byte_buffer
Memset the byte_buffer will screw up the byte_buffer.
Also limit memset to OUTPUT only as INPUT doesn't require the it when
closing the device. Remove the buf_reset as we already has buf_reset
when configuring the device so it's redundant to do so on close.

BUG=b:224800294
TEST=build,deploy, and test HFP recording while playback.

Change-Id: Id38dffaf7574eafcf5fb11c3a33f34ab6b8835aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3527814
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>