CRAS: iodev_list - reconnect pinned stream should check dev open state
remove_then_reconnect_stream function has a potential issue that
although the pinned stream exists the corresponding pinned iodev
may not be in open state.

The possible scenario for that is the iodev encountered EIO and
got closed in the error path earlier. Fixing this problem by
always check and init the iodev before reconnecting the stream.

BUG=b:223056958
TEST=This is for fixing a crash that happens in a rare code
path. Expect to see the crash disappear after this commit.

Change-Id: I29b30d12e325034447c794e1d0bd9d5341688096
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3531808
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>