CRAS: deprecate cras_bt_io_profile_to_log
The current implementation that decides the BT_DEVICE_PROFILE casts
cras_ionode to bt_node, which is only guaranteed to work on the legacy
BlueZ stack. The result is undefined if it is run under Floss.

This CL replaces the erroneous implementation by checking btflags,
which is accessible in any typical cras_ionode.

BUG=b:227120731
TEST=FEATURES=test emerge-${BOARD} adhd

Change-Id: I0917fd319f97781be75b4c10c20c7e6d50ee7fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3552005
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>