CRAS: deprecate usage of cras_bt_device_profile in bt_io_manager
We plan to make bt_io_manager an abstraction of profile switching
mechanisms that is agnostic of the BlueZ-specific details which
live in `cras_bt_device.c`.

In this CL, we replace all usages of the enum specific to
`cras_bt_device.c`, `cras_bt_device_profile`, with the more general
flags, `CRAS_BT_FLAGS`, to store information about whether the profile
is `A2DP` or `HFP`.

Since the old `profile` still lives in certain parts of the code, we
rename the replaced variables as `btflag` to avoid causing confusion.

BUG=b:227121218
TEST=FEATURES=test USE=asan emerge-${BOARD} adhd

Change-Id: Iac8518be990897cfbdcf9d571a3b0d0962639a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3581461
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>