sound_card_init: use VPD value when temperature is in invalid range
The problem happens when
1. there is no previous calibration and
2. the ambient temperature is invalid for boot time calibration,
at this time we should use VPD value.

This approach may induce problem when user replace the Chromebook internal
speaker to another speaker that is not the same model. Considering the
rarity of this case, this approach should be taken to avoid low volume
when calibration failed.

BUG=b:225754857
TEST=deploy on DUT and test if VPD value is used in invalid ambient temp

Change-Id: I199ab7f08d9ba39587d335998eff3dbc5014489d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3581466
Auto-Submit: Terry Cheong <htcheong@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>