Floss: Remove static hfp and a2dp_manager
Currently, we keep each of the cras_hfp/a2dp_manager in a static global
pointer. And, we also have the fl_media to control their creation and
destroy. This brings us limited benefits but complex the codes. By
removing these pointers, we can simplify the codes and also make it
flexible for the potential future requirements for multiple a2dp or hfp
connections.

BUG=b:230033979,b:225259380
TEST=unittest with asan, deploy and verify common use cases manually.

Change-Id: Id36ca5c15856a860063b897bc86ce877382e940d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3601527
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>