CRAS: Fix time-lag when attaching new streams
On reworking the implementation of `stream_added_cb` to fix a UAF bug,
we bumped into an issue where the time-lag between attaching a new
stream to the fallback and attaching to normal devices would break some
assumptions from the Android stack.

In this CL, we place the fallback enabling to right before reopening
device, and disabling if possible at the end of the call. The idea is to
avoid attaching a new stream unless upon reopening a device (which makes
it fine and the reason to do so).

BUG=b:232019938
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=tast run $DUT arc.AudioValidity.record_vm

Change-Id: I0c615eebeb84002fd56f406d3e05548035328dc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3644597
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>