CRAS: Fix `stream_list` before calls to `stream_(add|rm)_cb` in unittest
To add more extensive tests in `iodev_list_unittest.cc`, we have spotted
a few issues in the function wrappers causing the tests to behave in
ways that do not meet our intentions.

`stream_added_cb` and `stream_rm_cb` both expect the stream of concern
be updated in the `stream_list` that can be accessed by
`stream_list_get()`. The corresponding wrapper in
`iodev_list_unittest.cc` returns its own static list
`stream_list_get_ret`, and there are a few parts of the code forgetting
to commit the updates to that before invoking the callbacks.

BUG=b:231389631
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I9c989b45c144874a17b01f7e594bbd1704280493
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3627229
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>