CRAS: De-accumulate `EXPECT_EQ`s on number of functions calls
On reworking the implementation of certain functions in
`cras_iodev_list.c` (e.g., `stream_added_cb`), we found that we should
not assume that the number of calls to certain functions (e.g.,
`cras_iodev_open`, `audio_thread_add_open_dev`, `enable_device`, etc.)
is fixed. Otherwise, we will have to modify large parts of the unittest
whenever the number of calls to those functions changes.

On top of that, some of the `EXPECT_EQ` on the number of calls to those
functions are not bounded to their minimal scope. In other words, they
check the "accumulated" number of calls whenever "suitable". This can
cause troubles in two ways: (1) if the number of calls is changed in
between, all `EXPECT_EQ`s after that must be re-examined (2) it may be
"easier" to pass the tests when something is actually wrong, due to the
number being accumulated, especially as we are going to introduce some
`EXPECT_LE`s where we only assert the lower-bound of number of calls.

In this CL, we push each of such checks to its minimal boundary, so that
each of these `EXPECT_EQ` comparison work on the "difference" of number
of calls in between.

BUG=b:231389631
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: Ibec1e47be51464fbfb8b4ed8eb6a6fcdc12a56de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3626972
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>