CRAS: fix software_volume_needed in bt_io
`software_volume_needed` should be the opposite (boolean) value of
`use_hardware_volume`, but this logic got inverted in previous CLs. We
revert this in this CL.

Along with this bug, we found that `software_volume_needed` is not
always synced to the desired value due to it having multiple copies over
`bt_io`, `active_profile_dev`, and `active_node`. We enforce that
related read/writes in the domain of BT be on `bt_io` (i.e., the `iodev`
of this class that is inserted to the `iodev_list`).

BUG=b:232252034, b:232427319
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=BT Headphones are audible on lower volumes after hardware control

Change-Id: Ia4f991675821ca3acf4fe0dae337345bc8b1d66e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3653713
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>