bazel: Do not symlink descendants
This is the alternative of CL:3657216 which does not generates
symlink loops in the filesystem. Symlink loops turned out to be
problematic for oss-fuzz's coverage builder as source files are
copied with symlinks dereferenced.

Multiple invocations of pkg-config might return
include directories where one is a an ancestor of another.
Such as: "-I/a/b" and "-I/a/b/c".
In this case the symlink of system_libs/a/b/c -> /a/b/c
will fail since system_libs/a/b is already created.

This CL:
1. skips symlinking a include directory if its ancestor
   is also a include directory.
2. Use a per-library root.

Generated symlink tree:
pkg_config
├── libchrome
│   └── build
│       └── hatch
│           └── usr
│               ├── include -> /build/hatch/usr/include
│               └── lib64
│                   ├── dbus-1.0
│                   │   └── include -> /build/hatch/usr/lib64/dbus-1.0/include
│                   └── glib-2.0
│                       └── include -> /build/hatch/usr/lib64/glib-2.0/include
└── libcros_config
    └── build
        └── hatch
            └── usr
                └── include
                    └── chromeos -> /build/hatch/usr/include/chromeos

BUG=b:233361473
BUG=b:251701189
TEST=FEATURES=test emerge-volteer adhd
TEST=python3 infra/helper.py build_fuzzers --clean --sanitizer coverage cras /path/to/cras

Change-Id: Idfcf60cb76bd4d90d4ec1f43c1c61173331bc429
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3660780
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>