audio_processor: Add CheckShape
`CheckShape` returns the input unmodified if the input matches the specified
shape. Otherwise an error is returned.

The intended use of CheckShape is to terminate the last block of a
of a generating processor early, which might be incomplete.
For example, if a `WavSource` is configured with block size 480
and reads from a WAVE file of 1000 frames, it will generate results
of [Ok(480), Ok(480), Err] respectively.

BUG=b:232742804
TEST=cargo build && cargo test

Change-Id: I733f9bffc589480ecdf3d30d43d89cb40042d47a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3787816
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>