audio_test_gui: Fix nodes missing in webpage view.
Currently, due to cras_test_client output format changes, nodes are
missing when viewed through the webpage, even when they are plugged in.
Also, the node name may be incomplete if there are spaces in the name.

Change the parsing of cras_test_client --dump_s output to ensure that
nodes appear in the webpage view correctly, and that they have correct
names.

BUG=b:237339691
TEST=python audio_test_gui.py, connect to localhost:8080 and compare
     devices shown with cras_test_client --dump_s output.

Change-Id: I3c87e6f2f4e2febacd889d6173de9bc019d696e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3735986
Commit-Queue: Baili Deng <bailideng@google.com>
Tested-by: Baili Deng <bailideng@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>