CRAS: fix btflags propagation for runtime metrics
`btflags` is an attribute that holds BT-related information about an
ionode, including the BT stack, offloading/, and A2DP/HFP, the current
use of which is to reflect the correct media type to UMA metrics.

`btflag` is a similar attribute for internal use that only stores
information about A2DP/HFP and was correctly reflected.

In this CL, we make sure that `btflags` is propagated as `btflag` is.

BUG=None
TEST=USE=asan FEATURES=test emerge-${BOARD} adhd
TEST=Deploy on DUT, verify the log shows correct type emitted to UMA

Change-Id: I2addf64e2a1075af0aa1fe5f8c85cb78b5f15cd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3740239
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>