CRAS: block StartAudioRequest until audio state changed
When the DBus method `StartAudioRequest` is invoked from CRAS, it is
immediately returned regardless of whether the audio state is ready in
Floss.

We need a mechanism to wait until the audio state is ready, and our
decision is to expose a new DBus method `GetA2dpAudioStarted` for CRAS
to poll immediately after the request.

BUG=b:236769419
TEST=Apply full patch series, connect BT headset and play Youtube.
Verify pausing and resuming clips are well-synced.

Cq-Depend: chromium:3760759
Change-Id: If07eef5479fb3f453dec5b23c48f9c182213d86d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3747464
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>