cras: Remap microphone UI gain range
Currently, the microphone dB gain ranges from -40dB to 20dB. This
causes sounds recorded from the microphone to have extremely low volume
when the UI slider is turned down all the way. The volume is so low,
that the user could easily mistake the microphone for being off or
broken.

Modify the negative db gain range to -20dB to 0dB. This range maps to 0
to 50 on the UI gain range.

-20dB corresponds to a 0.1 in ampltiude ratio, compared to 0.01 in
amplitude ratio for -40dB. From testing, this causes the recorded voice
to sound quieter, but not too quiet as to be easily missed.

BUG=b:238280339
TEST=Open Google Meet and test the volume manually while lowering the UI
microphone gain slider. The headphone volume is set to allow clear
hearing at medium volume when the microphone gain is set at 50.

Change-Id: I3588832c0d58845da7d0baba4d172007bacde9a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3752507
Tested-by: Baili Deng <bailideng@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Baili Deng <bailideng@google.com>