CRAS: apm_reverse - Add mutex to protect fbuf
cra_apm_reverse module runs in the context of a cras_dsp_pipeline.
We have a mutex in cras_dsp_pipeline to protect it from race by
its configure and process calls.
However this pipeline mutex alone is not enough to handle
apm_reverse use case, because when output iodev switches that
is two different mutexes guarding two calls on the same object.

We fix the issue by adding a mutex to cras_apm_reverse_module,
guarding its configure() and run() calls. This is considered of
minimal impact to audio thread because reverse_data_configure
simply reallocates a buffer so it executes fast.

BUG=b:238273052
TEST=Run audio_CrasDevSwitchStress on zork and pass

Change-Id: Ie03921753ecec1c58fd0d40147df5f6959d51554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3752282
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>