CRAS: add sco-offloading HFP for Floss
In this CL, we implement sco-offload logic to Floss by modifying the
existing hfp_alsa_io (sco-offload for BlueZ) to share common constructs.

Currently, there are still pending CLs to be merged in Floss and the
kernel so that this could take off. However, we force `is_sco_pcm_used`
to be `false` so that we could land the parts in CRAS without affecting
current users.

BUG=b:237373343
BUG=b:233322784
TEST=USE=asan FEATURES=test emerge-${BOARD} adhd

Change-Id: I20b1eafb88180b0fa02e494972047fc6c079100f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3757040
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>