libcras: Ignore capture data with incomplete frame size
When we receive an incomplete data while capturing, drop the message
and wait for the next one.

This happens when the server overrun and overwrite partial data to the
buffer, then client tried to read the data from an older message.

Add this behavior to rust lib as it is already in the C client.

BUG=b:236091162
TEST=1. Simulate timing issue by adding `sleep()` to crosvm code
     2. Test voice recorder app in ARCVM
     3. Check that there is no "Bytes written {} != period_bytes {}"
        error

Change-Id: I5f0eff63be017ca01384acf4bc2228f169f0c000
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3760313
Tested-by: Pattara Teerapong <pteerapong@chromium.org>
Commit-Queue: Pattara Teerapong <pteerapong@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>