CRAS: fix improper timing of Floss SCO fd connection
Currently a constant value of sleep duration (1s) is used to wait for
datapath to be ready in Floss.

However, such timing is not safe and could lead to multiple issues,
whether due to it being too short or too long.

In this CL we fix that by polling the HFP connection state and blocking
the function (connecting to the fd) until as soon as it's ready.

BUG=b:241050004
TEST=Apply full patch, verify CVSD/mSBC HFP consistently starts

Cq-Depend: chromium:3820944
Change-Id: I208178e5ae824706a9d7414ee78aa2d6822f89b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3802011
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>