CRAS: nullify active_fm in fl_media_destroy
Since CL:3788001, active_fm, the static variable that holds information about
the Floss media interface, does not get nullified in fl_media_destroy
due to pass-by-value.

In this CL we change that to pass-by-reference so that fl_media gets
nullified as intended and in that way double-free would not happen.

BUG=b:241354214
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=Deploy and verify CRAS doesn't crash when btadapter crashes

Change-Id: If43e5a21e9b6adb7b0b0c7e756775c8089e691c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3812709
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>