sound_card_init: Add fstatfs(64) syscalls
While attempting to benchmark the scudo allocator on ChromeOS,
I encountered seccomp failures in sound_card_init. I don't think
these failures are related to scudo itself, and are simply
missing seccomp entries in the sound_card_init. fstatfs was
already present in the arm32 policy (interestingly, not fstatfs64),
so this adds it to the remaining seccomp files for parity.

For clarity:
* amd64 was missing fstatfs, and has been updated here.
* arm32 had fstatfs, but was missing fstatfs64 which
  is needed for 32bit. fstatfs64 has been added to future proof.
* arm64 was missing fstatfs, and has been updated.

BUG=None
TEST=CQ

Change-Id: I02b6a1136b72bfe53cc26f904787f8d75253354e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3832445
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>