CRAS: reconnect when HFP hangs up in Floss
Some HF devices (e.g., WH-1000XM*) resets SCO connection when removed
from ears.

With the BlueZ stack, we used to close the connection and wait until the
user manually starts another SCO connection to get back to a normal
state. However, in Floss we currently leave the stale connection with
the callback in the audio thread, and run into a futile busyloop.

In this CL, we attempt to reconnect to the SCO, so that the user will
hardly notice the disconnection.

BUG=b:241527929
TEST=Deploy and verify it reconnects as soon as WH* is removed from ears

Change-Id: I0fd3bd9a92a4efab59867d41dbbac5e74155f882
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3837535
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>