CRAS: reconnect when HFP hangs up in BlueZ
Some HF devices (e.g., WH-1000XM*) reset SCO connection when removed
from ears.

With the BlueZ stack, we used to close the connection and wait until the
user manually starts another SCO connection to get back to a normal
state.

In this CL, we attempt to reconnect to the SCO, so that the user will
hardly notice the disconnection.

BUG=b:241527929
TEST=Deploy and verify it reconnects as soon as WH* is removed from ears

Change-Id: I39bd4f23d8515e7c4cb612a0dcb797f4bbe34605
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3840988
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>