CRAS: add unittest for profile switch flag
In previous work (CL:3830145) we identified issues around the
profile-switch event that we had to introduce a new flag to prevent
unexpected second-profile-switch event from emerging when an Hfp iodev
is opened before it is resolved.

Also, in CL:3837535 we start to allow profile-switching from Hfp to Hfp
as a means to restart, in which case we want to prevent it from spawning
another profile-switch when it is closed.

In this CL, we check that opening/closing Hfp iodev will not spawn the
new event, and that the flag is cleared when the profile-switch has
passed through the critical section.

BUG=b:242665612
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I31037d146bbb776ec665c7d5da1dd4100e3ffe82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3853738
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>