CRAS: move sco-offloading connection to open_dev
In b/241202081 we learned that there are cases where the format can be
decided only after SCO connection is established.

We have to move the connection request into open_dev because it is
expected that the format is determined by the end of it. This
solution is already applied in hfp_iodev (BlueZ non-offload) and
fl_pcm_iodev (Floss non-offload).

In this CL, we apply the same change onto hfp_alsa_iodev (offload).

BUG=b:237373343
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: Ie19ba17e6393e875d33433df211c8a82fb892a17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3858325
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>