CRAS: explictly set_active_device to refresh session in Floss
In Floss, some cleanups are done in `end_session`, which is explicitly
called when `set_active_device` is invoked with a null-address argument.

In this CL, we clear the session each time the `active` state has
changed, and only set the device to `active` when necessary.

BUG=b:240641968
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=Apply all patches and verify various ways of A2DP restart

Change-Id: Ibb9e4019140b4033562386e445396057f22cbd85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3874073
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>