CRAS: reconnect when HFP is disconnected unexpectedly in Floss
In CL:3837535 we have addressed the issue where the audio thread could
run into busy loop due to the HF device asking for reconnection (which
was ignored by CRAS).

However, we did so in the SCO read-write callback that is only within
our reach when SCO is not offloaded.

In this CL, we create a new callback for Floss to notify of audio
disconnection, since otherwise there is no simple and general way to
retrieve that information, and use that instead to decide whether to
reconnect.

BUG=b:241527929
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=Verify HFP reconnects whether offloaded or not

Cq-Depend: chromium:3905665
Change-Id: I276f71867389dc67a6ce8ca8926da5e626236d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3881570
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>