ucm: Use geti for value lookups
We are using loops to iterate over lists to lookup statuses, this uses a
bunch of string logic and comparisons. I am not sure if this is due to
lack of understanding of the UCM API or legacy API functions, but it
should be updated to direct accesses.

BUG=None
TEST=None

Change-Id: Id344d06dcd0179c2d6ceeb7b069fdbc9574ac4e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3908212
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>