CRAS: shorten timeout for A2DP start request in Floss
We added a blocking mechanism (to wait for audio state ready) in Floss
to resolve the A/V sync issue in b/236769419.

By the time we discovered that due to some configuration
(max_sniff_period being 500ms), we had to wait for 1-2s frequently and
at worst cases 3-4s.

However, from the UMA metrics (see bug for details) we can see that >99%
of configure time was less than 1s, 100% less than 1.5s.

In this CL, we change the timeout to 2s.

BUG=b:239370946
BUG=b:245657624
TEST=Build and verify

Change-Id: I4bc0960b4f1ccb63451a765b8de43487c274a371
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3899781
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>