CRAS: Add output control by soundcard type
Initially, we will add mixer_control to main_volume_control or output_control based on the mixer_control name.
The reason we do this is because on some older boards we need paths to enumerate auido nodes. But sometimes the mixer_control name of some USB devices belongs to main_volume_control.
This will cause what should be added to output_control to be added to main_volume_control.
For all usb devices, they don't need to use the old path to create the auido node, so we need to create the mixer_control according to the sound card type.
If the sound card type is interal then keep the original logic, but if the sound card type is usb then create all output controls to output_control

BUG=b:246268778
TEST=FEATURES=test USE=asan emerge-amd64-generic adhd and verify Jabra Speak 510 SME work.

Change-Id: Iddc6db5485c4ecb50efda235e9fc81d36f4bf55c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3903805
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Ching Yun Chang <whalechang@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Ching Yun Chang <whalechang@google.com>