CRAS: make get_sco_pcm_iodev return unique value
Since both `alsa_io` and `hfp_alsa_io` can have the flag
`CRAS_BT_FLAG_SCO_OFFLOAD`, the return value of this function may be
ambiguous.

In this CL, we enforce the return value to be the `alsa_io` which is
unique.

Note that this change has no effect at the moment.

BUG=None
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I06b03673dcecb0786fea658936647379361a90ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3906796
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>