CRAS: Secure encoded mSBC packet padding
It is specified in the HFP spec that the padding byte should be zero.
However, we may send out packets with non-zero byte as padding and lead
undefined behavior that varied on HF's implementation due to the use of
malloc and the write of 57 bytes instead of 58 bytes for the encoded
mSBC frame.

BUG=b:248619908
TEST=Build, deploy and verify

Change-Id: I7da9d1134afe3725788e6d8de91ea133315d93dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3916698
Commit-Queue: Jeremy Wu <jrwu@google.com>
Auto-Submit: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Jeremy Wu <jrwu@google.com>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>