alsa_benchmark: Fix ucm-suffix path
cros_config.GetString is expecting the second parameter to be a property
string which it ultimately appends to a file path. file_path.cc Append has
DCHECK(!IsPathAbsolute(appended)), which we are triggering when running this
benchmark on a test image. Remove the leading / to avoid this.

BUG=b:238411692
TEST=cras_bench --benchmark_format=json --benchmark_filter=BM_Alsa

Change-Id: I944f4adb016216733170655e860637397c12320e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3951353
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>