CRAS: return error when blocking call has no response in Floss
Currently, in `floss_media_block_until_started`, since the default
return code is 0, it will return 0 even when it hasn't started.

CRAS is relying on later checks with the socket availability, but that
could cause other racing issues (codec mismatch).

In this CL, we set the return code to be an error code so that it
will fail right away when the timeout is over and potentially schedule a
retry.

BUG=b:255706576
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I37c4cf507157bbb637557710f5ba68840d13a6d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3985506
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>