CRAS: iodev - Remove all pre-processing capture mute
In commit c7f19aa2 we added two layers of software mute to the
capture path - before and after APM processing. The 2nd mute
is to make sure any artifacts generated by processing got muted
as well.
This change removes the 1st SW mute and the alsa mixer mute
as the preparation for speak on mute feature, in which we rely
on APM processing the real captured content to tell if there's
voice activity when system input is muted.

BUG=b:255935803
TEST=Apply full patch set, launch Google Meet, mute input from UI
and verify dBus signal `dbus-monitor --system | grep Speak`

Change-Id: Ic52bcefd132014705e68805236124477b3405357
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3990485
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Li-Yu Yu <aaronyu@google.com>