CRAS: warn instead of error for init failure in profile-switch
With the current design, all attempts to open a bt device during
profile-switch is meant to be failed.

However, those failures are caught as errors from the general
perspective, and we have caused much misunderstanding with these logs.

In this CL, we catch the |EAGAIN| return code, change the error to
warning, and make a special message to hint about profile-switch. This
relies on the fact that when this code is shown, it is almost surely due
to profile-switch.

BUG=b:255707592
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I623852d311e15ccb4c76b315e19a4fb010e800df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4019841
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>