CRAS: makes sr bt able to be force enabled
Background:
In autotest, we have a test for the sr bt feature, which tries to start
chrome with the feature flag enabled to enable the feature in cras.
However, the step of starting chrome will fail in stainless but pass
locally(see the bug below for details.)

Change:
cras_sr_bt_can_be_enabled is the location we check the feature flag
status. This change makes it checkes the system state, force_sr_bt_enabled,
first: if it's force enabled, then the check of the feature flag is skipped.

This is a workaround and only for testing purpose.

BUG=b:251837712
TEST=FEATURES="test noclean" USE="asan cras-ml" emerge-amd64-generic -j adhd

Change-Id: I975112fa5a58e37fca8ea086183fb2720d265c99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4025412
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Cranel W <cranelw@chromium.org>
Tested-by: Cranel W <cranelw@chromium.org>