CRAS: destroy audiogate/endpoint after reset in BlueZ
Currently, when |cras_bt_stop_bluez| is called, it destroys the
audiogate/endpoint before they are reset.

In other words, before deregistering and detaching from its associating
members, the audiogate/endpoint is destroyed, leading to UAF.

Notice how |cras_bt_reset| relies on |profiles| and |endpoints| be
intact to proceed with the correct order of cleanup.

In this CL, we move the destroy to come after reset, as |profiles| and
|endpoints| will otherwise not be cleared as intended. Also note that
these two objects are meant to have static lifetimes and thus is safe
to be referred by after |cras_bt_reset|.

BUG=b:256851566, b:232043189, b:259345358
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I7664477e5a11008716889f8ba4485a9638ef2e25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4040044
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Fernando Perea <pirunchy20@gmail.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>