CRAS: make HFP auto reconnection a warning message
We have landed a slightly aggressive patch to deal with unexpected
HF-side SCO disconnection in CL:3881570.

When some feedback reports show multiple reconnections in a short
interval of time, there is no way to tell if it was the user's action or
if it was the result of this mechanism.

In this CL, we change the log level to warning since it is something
worthy of seeing in FRs.

BUG=b:260618573
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I441baeab8fa7b5337cc15edaf238061d56e08a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4063539
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>