CRAS: generalize metafunction for blocking call to Floss
We expect to have more complicated arguments and predicates to face with
dbus methods in Floss that require the blocking/retry mechanism.

In this CL, we generalize the function so that it could take arbitrary
arguments and judge conditions based on custom functions.

BUG=b:259420892
TEST=Verify A2DP and HFP are functioning

Change-Id: Ia1d130d1edf9634a80148055bb5b20d0e8f7eb30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4082424
Auto-Submit: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>