Restore code to unmute capture ALSA mixer
In CL:3990485, we removed the capture ALSA mixer mute and relied on mute to be applied after APM processing. The removed code was also responsible for unmuting the mixer.
There are drivers where the mixer is muted by default, combined with the
removal results in capture to always return zeros in those devices.

This commit restores part of the code to ensure that the capture ALSA mixer is not muted on these devices.

BUG=b:255935803
BUG=b:261971615
TEST=audio.CrasRecordQuality test on reven x1cg8

Change-Id: Icd204e398fff4e54ef834bbd248b8a859074a62c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4095957
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>