CRAS: retry |start_audio_request| until accepted in Floss
Requests to start the audio are non-blocking and are not guaranteed to
be fulfilled. Currently, Floss exposes |get_x_audio_connected| to allow
the audio stack to query the audio state and decide if it should retry
or give up on the request in case it takes too long.

However, requests could be rejected immediately (e.g., due to Floss
being temporarily busy with other requests), in which case CRAS needs to
know so that it doesn't continue querying the audio state.

In this CL, we block on the request status for A2DP. Also, we rename
some of the functions and signatures to better represent their purpose.

BUG=b:259420892, b:245657624
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Cq-Depend: chromium:4130524
Change-Id: I8a00150b5a6b24a006a33df11083a28c7c570d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4040048
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>