CRAS: bounds check on number of channels in dsp
We have no bounds check on the number of channels, but there are exactly
|MAX_EXT_DSP_PORTS| of ports (i.e., buffers) allocated in the data
structure held by the DSP.

If a device claims to have more channels than the limit, it will
relentlessly access the memory beyond the bounds of that region and
result in crashes.

In this CL, we add some bound checks to the loops that attempt to
read from illegal offsets, by simply ignoring them for now.

BUG=b:263226203
TEST=FEATURES=test USE=asan emerge-$BOARD adhd
TEST=Verify a high-channel device doesn't crash on capture/playback

Change-Id: Ifbefa98d454ce3279646d628a951be306fc88537
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4113399
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>